TelemetryContext should be a struct
